refactor(workspace): extract tests via #[path]
Phase 0 PR-3 of the architecture decoupling roadmap (#1082, #1083).
Same mechanical pattern: move workspace.rs's `#[cfg(test)] mod tests`
block into a separate file via `#[cfg(test)] #[path = "..."] mod tests;`.

workspace.rs is the largest single source file in the crate at 2950
lines. The tests block is the largest extractable test mod still
embedded in production code.

Net effect:
  workspace.rs:        2950 -> 1439 lines (-1511)
  workspace_tests.rs:   NEW   1511 lines

Zero logic changes. The new file uses `use super::*;` so it sees the
workspace module unchanged. No public API surface change.

Verification:
- cargo test -p harness-server --lib -- workspace:
    54/54 passed (covers workspace::tests submodule plus cross-module
    consumers periodic_retry::tests, http::tests::get_task_exposes_*,
    handlers::overview::tests::worktrees_used_*)
- RUSTFLAGS=-Dwarnings cargo check --workspace --all-targets: clean
- cargo fmt --all -- --check: clean

workspace.rs is still over the U-16 800-line ceiling (1439 lines
remaining). The remaining body is workspace lifecycle + reconciliation
+ registry; future Phase 1 cohesive splits can target it. This PR is
the prerequisite — without test extraction, edits to workspace.rs
were pre-blocked by the file-size guard.
